+++ module-template/tests/container.go
@@ -1,0 +1,95 @@
+package main
+
+import (
+	"context"
+	"strings"
+
+	"github.com/Excoriate/daggerverse/module-template/tests/internal/dagger"
+)
+
+// TestUbuntuBase tests that the target module is based on the Ubuntu 22.04 image.
+//
+// This function verifies that the target module is configured appropriately to use the base Ubuntu 22.04 image.
+// It runs a command to get the OS version and confirms it matches "Ubuntu 22.04".
+//
+// Arguments:
+// - ctx (context.Context): The context for the test execution.
+//
+// Returns:
+//   - error: Returns an error if the Ubuntu image is not used or if the output is not as expected.
+func (m *Tests) TestUbuntuBase(ctx context.Context) error {
+	targetModule := dag.
+		ModuleTemplate().
+		BaseUbuntu(dagger.ModuleTemplateBaseUbuntuOpts{Version: "22.04"})
+
+	out, err := targetModule.Ctr().
+		WithExec([]string{"grep", "^ID=ubuntu$", "/etc/os-release"}).
+		Stdout(ctx)
+
+	if err != nil {
+		return WrapError(err, "failed to get Ubuntu image")
+	}
+
+	if !strings.Contains(out, "ID=ubuntu") {
+		return WrapErrorf(err, "expected Ubuntu 22.04 or ID=ubuntu, got %s", out)
+	}
+
+	return nil
+}
+
+// TestAlpineBase tests that the target module is based on the Alpine Linux v3.17.3 image.
+//
+// This function verifies that the target module is configured appropriately to use the base Alpine Linux v3.17.3 image.
+// It runs a command to get the OS version and confirms it matches "Alpine Linux v3.17.3".
+//
+// Arguments:
+// - ctx (context.Context): The context for the test execution.
+//
+// Returns:
+//   - error: Returns an error if the Alpine image is not used or if the output is not as expected.
+func (m *Tests) TestAlpineBase(ctx context.Context) error {
+	targetModule := dag.ModuleTemplate().
+		BaseAlpine(dagger.ModuleTemplateBaseAlpineOpts{Version: "3.17.3"})
+
+	out, err := targetModule.Ctr().WithExec([]string{"cat", "/etc/os-release"}).Stdout(ctx)
+	if err != nil {
+		return WrapError(err, "failed to get Alpine image")
+	}
+
+	// Adjust the conditions to match the actual output.
+	if !strings.Contains(out, "Alpine Linux") || !strings.Contains(out, "VERSION_ID=3.17.3") {
+		return WrapErrorf(err, "expected Alpine Linux VERSION_ID=3.17.3, got %s", out)
+	}
+
+	return nil
+}
+
+// TestBusyBoxBase tests that the target module is based on the BusyBox v1.35.0 image.
+//
+// This function verifies that the target module is configured appropriately to use the base BusyBox v1.35.0 image.
+// It runs a command to get the OS version and confirms it matches "BusyBox v1.35.0".
+//
+// Arguments:
+// - ctx (context.Context): The context for the test execution.
+//
+// Returns:
+//   - error: Returns an error if the BusyBox image is not used or if the output is not as expected.
+func (m *Tests) TestBusyBoxBase(ctx context.Context) error {
+	targetModule := dag.
+		ModuleTemplate().
+		BaseBusyBox(dagger.ModuleTemplateBaseBusyBoxOpts{Version: "1.35.0"})
+
+	out, err := targetModule.Ctr().
+		WithExec([]string{"busybox", "--help"}).
+		Stdout(ctx)
+
+	if err != nil {
+		return WrapError(err, "failed to get BusyBox image")
+	}
+
+	if !strings.Contains(out, "v1.35.0") {
+		return WrapErrorf(err, "expected BusyBox v1.35.0, got %s", out)
+	}
+
+	return nil
+}
